Expand the response status header regex to reduce false positives (#1)
Currently it can match something like "Date: Wed, 17 Feb 2021 18:52:55 GMT", so
we anchor it to start with "HTTP/<non-whitespace><whitespace>202" to reduce that
possibility.

This isn't a perfect fix as I imagine the implementation for parsing the status
code in other frameworks, in other languages, is more robust. However PHP has
nothing along these lines in the standard library as far as I found find, so
this seemed like a decent compromise.